renamed url route translate to /translate/:id/edit
